Moved interfaces and throughput info structures to api module
