+++ datastream-server-api/src/main/java/com/linkedin/datastream/server/ClusterThroughputInfo.java
