+++ datastream-server-api/src/main/java/com/linkedin/datastream/server/PartitionThroughputInfo.java
@@ -7,19 +7,21 @@ package com.linkedin.datastream.server;
 public class PartitionThroughputInfo {
   private final int _bytesInRate;
   private final int _messagesInRate;
+  private final String _partitionName;
 
   /**
    * Creates an instance of {@link PartitionThroughputInfo}
    * @param bytesInRate Bytes in rate for the partition
    * @param messagesInRate Messages in rate for the partition
    */
-  public PartitionThroughputInfo(int bytesInRate, int messagesInRate) {
+  public PartitionThroughputInfo(int bytesInRate, int messagesInRate, String partitionName) {
     _bytesInRate = bytesInRate;
     _messagesInRate = messagesInRate;
+    _partitionName = partitionName;
   }
 
   /**
-   * Gets the bytes in rate
+   * Gets the bytes in rate (in KB)
    */
   public int getBytesInRate() {
     return _bytesInRate;
@@ -31,4 +33,9 @@ public class PartitionThroughputInfo {
   public int getMessagesInRate() {
     return _messagesInRate;
   }
+
+  /**
+   * Gets the partition name
+   */
+  public String getPartitionname() { return _partitionName; }
 }
